MQE: improve performance of subqueries when the parent query is a range query with many steps (#9719)
* Introduce ring buffer view types

* Use new ring buffer view methods everywhere

* Introduce pointer type for ring buffer view

* Make use of the fact that the head slice from `UnsafePoints` is always populated

This also avoids unnecessarily calling `accumulate()` on the first
point in `floatRate()` for `rate`.

* Improve performance of `UnsafePoints` by using subtraction rather than remainder operation

* Don't pass entire RangeVectorStepData to `floatRate` and `histogramRate`

* Improve performance of `UnsafePoints` if view is empty

* Split `ViewUntil`

* Use pointer for `RangeVectorStepData` to avoid copying fields

* Replace use of remainder with bit masking

* Add changelog entry

# Conflicts:
#	CHANGELOG.md

* Apply suggestions from code review


* Address PR feedback: refactor tests

* Address PR feedback: add tests for reusing view

* Address PR feedback: add comment about possible future optimisation

* Address PR feedback: ensure view is reused in tests

---------

+++ pkg/util/pool/bucketed_pool.go
@@ -21,7 +21,7 @@ type BucketedPool[T ~[]E, E any] struct {
 
 // NewBucketedPool returns a new BucketedPool with size buckets for minSize to maxSize
 // increasing by the given factor.
-func NewBucketedPool[T ~[]E, E any](minSize, maxSize int, factor float64, makeFunc func(int) T) *BucketedPool[T, E] {
+func NewBucketedPool[T ~[]E, E any](minSize, maxSize int, factor int, makeFunc func(int) T) *BucketedPool[T, E] {
 	if minSize < 1 {
 		panic("invalid minimum pool size")
 	}
@@ -34,7 +34,7 @@ func NewBucketedPool[T ~[]E, E any](minSize, maxSize int, factor float64, makeFu
 
 	var sizes []int
 
-	for s := minSize; s <= maxSize; s = int(float64(s) * factor) {
+	for s := minSize; s <= maxSize; s = s * factor {
 		sizes = append(sizes, s)
 	}
